requirements.txt
beautifulsoup4
requests
pdfminer.six
sqlalchemy
apscheduler
python-slugify
fastapi
uvicorn
pytest
vcrpy
pytest-mock
